ui/requirements.txt
google-cloud-aiplatform>=1.25.0
google-api-python-client>=2.87.0
google-cloud-core>=2.3.2
graphdatascience
retry==0.9.2
langchain>=0.0.216
streamlit>=1.25.0
streamlit-chat>=0.1.1
streamlit-chat-media>=0.0.4
numpy
pandas
webvtt-to-json
webvtt-py
requests
plotly==5.15.0
altair<5